fix(schematics): table schematic not expanding full width (#13234)
Currently the `table` schematic generates a `<table>` element that does not expand to the available width. By default, we should expand to the full width. Similarly to every table example on our docs. This makes it easier to play with the table and also gives people a better first glance of a table. Also currently it doesn't look that good (in perspective of layout) because the paginator expands to full width, but the table not.